reverted MClust 3.5 LoadSpikes() back to 32bit
should really have autodetect somehow

+++ code-matlab/toolboxes/MClust-3.5/Utilities/IOUtils/LoadSpikes.m
@@ -48,7 +48,7 @@ for iF = 1:nFiles
 		end
 		
 		ReadHeader(tfp);    
-		S{iF} = fread(tfp,inf,'uint64');	%read as 32 bit ints
+		S{iF} = fread(tfp,inf,'uint32');	%read as 32 bit ints
 		
 		% Set appropriate time units.
 		switch (tsflag)
